MyTime redifine to work more accurate (with float type)

+++ mytime.h
@@ -4,19 +4,19 @@ class QString;
 class MyTime
 {
 private:
-    int m_days;
-    int m_hours;
-    int m_minutes;
+    float m_days;
+    float m_hours;
+    float m_minutes;
 
 public:
     MyTime();
-    MyTime(int days, int hours, int minutes);// (0д, -15ч, -5мин)
+    MyTime(float days, float hours, float minutes);// (0д, -15ч, -5мин)
 
     void adjust();                              //перевести, чтобы было правильно
 
-    int days() const {return m_days;}
-    int hours() const {return m_hours;}
-    int minutes() const {return m_minutes;}
+    float days() const {return m_days;}
+    float hours() const {return m_hours;}
+    float minutes() const {return m_minutes;}
 
 public:
     int toMinutes() const {return m_days * 24 * 60 + m_hours * 60 + m_minutes;}
@@ -28,8 +28,8 @@ public:
             return m_days * 24 + m_hours + 1;
     }
 
-    static MyTime timeFromHours(int hours);
-    static MyTime timeFromMinutes(int minutes);
+    static MyTime timeFromHours(float hours);
+    static MyTime timeFromMinutes(float minutes);
 
     MyTime operator+ (const MyTime t) const;
     MyTime operator- (const MyTime t) const;
